+++ docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Fixed non-playable character types not prefixing their internal `CharacterType` with the mod ID. This may break mods adding non-playable (enemy, story) characters.
+
 ## 1.16.0
 Released 27 April 2025.